requirements.txt
blinker==1.9.0
click==8.2.0
Flask==3.1.1
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
fpdf==1.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.0
MarkupSafe==3.0.2
numpy==1.26.4
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
scikit-learn==1.6.1
scikit-surprise==1.1.4
scipy==1.15.3
six==1.17.0
SQLAlchemy==2.0.41
surprise==0.1
threadpoolctl==3.6.0
typing_extensions==4.13.2
tzdata==2025.2
Werkzeug==3.1.3
